Move searchIncludes to SearchLayer
