+++ map/src/manager/SearchManager.js
@@ -1,7 +1,14 @@
 import { apiGet } from '../util/HttpApi';
 import {
+    BACKGROUND_TYPE_EXTENSION,
+    CATEGORY_NAME,
+    CATEGORY_TYPE,
+    COLOR_NAME_EXTENSION,
+    FINAL_POI_ICON_NAME,
+    ICON_KEY_NAME,
     MAIN_CATEGORY_KEY_NAME,
     POI_ID,
+    POI_NAME,
     WEB_POI_ADDITIONAL_CATEGORY,
     WEB_POI_FILTER_NAME,
 } from '../infoblock/components/wpt/WptTagsProvider';
@@ -9,11 +16,17 @@ import capitalize from 'lodash-es/capitalize';
 import { formattingPoiType } from './PoiManager';
 import { getFirstSubstring } from '../menu/search/search/SearchResultItem';
 import i18n from 'i18next';
-import { getObjIdSearch, SEARCH_ICON_MAP_LOCATION, typeIconMap } from '../map/layers/SearchLayer';
+import {
+    FAVORITE_HIT_GROUP_ID,
+    getObjIdSearch,
+    SEARCH_ICON_MAP_LOCATION,
+    searchTypeMap,
+    typeIconMap,
+} from '../map/layers/SearchLayer';
 import { DEFAULT_EXPLORE_POITYPES } from '../menu/search/SearchMenu';
 import { OBJECT_TYPE_POI, OBJECT_TYPE_CLOUD_TRACK, OBJECT_TYPE_FAVORITE } from '../context/AppContext';
 import { openFavoriteObj } from './FavoritesManager';
-import { openTrackOnMap } from './track/TracksManager';
+import { EMPTY_FILE_NAME, getGpxFiles, openTrackOnMap, prepareName } from './track/TracksManager';
 import { MAIN_URL_WITH_SLASH, SEARCH_URL, SEARCH_RESULT_URL } from './GlobalManager';
 import { buildSearchParamsFromQuery } from '../util/hooks/search/useSearchNav';
 
@@ -25,6 +38,84 @@ export const SEARCH_BRAND = 'brand';
 
 export function getIconByType(type) {
     return typeIconMap[type] ?? SEARCH_ICON_MAP_LOCATION;
+}
+
+function searchIncludes(text, query, collator) {
+    const textChars = [...String(text ?? '')];
+    const queryStr = String(query ?? '');
+
+    if (!queryStr) {
+        return true;
+    }
+
+    const queryLength = [...queryStr].length;
+
+    for (let i = 0; i <= textChars.length - queryLength; i++) {
+        const candidate = textChars.slice(i, i + queryLength).join('');
+
+        if (collator.compare(candidate, queryStr) === 0) {
+            return true;
+        }
+    }
+
+    return false;
+}
+
+export function searchFavoriteFeatures({ favorites, query, collator }) {
+    const q = String(query ?? '').trim();
+    if (!q || !favorites?.groups?.length || !favorites.mapObjs) {
+        return [];
+    }
+
+    return favorites.groups
+        .filter((group) => group?.id)
+        .flatMap((group) => {
+            const wpts = favorites.mapObjs[group.id]?.wpts;
+            if (!wpts?.length) return [];
+
+            return wpts
+                .filter(
+                    (wpt) =>
+                        wpt?.name &&
+                        wpt.lat != null &&
+                        wpt.lon != null &&
+                        (searchIncludes(wpt.name, q, collator) || searchIncludes(wpt.desc ?? '', q, collator))
+                )
+                .map((wpt) => ({
+                    type: 'Feature',
+                    geometry: {
+                        type: 'Point',
+                        coordinates: [wpt.lon, wpt.lat],
+                    },
+                    properties: {
+                        [CATEGORY_TYPE]: searchTypeMap.FAVORITE,
+                        [CATEGORY_NAME]: wpt.category,
+                        [POI_NAME]: wpt.name,
+                        [FAVORITE_HIT_GROUP_ID]: group.id,
+                        [ICON_KEY_NAME]: wpt.icon,
+                        [COLOR_NAME_EXTENSION]: wpt.color,
+                        [BACKGROUND_TYPE_EXTENSION]: wpt.background,
+                        [FINAL_POI_ICON_NAME]: wpt.icon,
+                        ...(wpt.address ? { address: wpt.address } : {}),
+                    },
+                }));
+        });
+}
+
+export function searchCloudTrackFeatures({ listFiles, query, collator }) {
+    const q = String(query ?? '').trim();
+    if (!q || !listFiles?.uniqueFiles) return [];
+
+    return getGpxFiles(listFiles)
+        .filter((f) => !f.name.endsWith(EMPTY_FILE_NAME) && searchIncludes(prepareName(f.name, true), q, collator))
+        .map((f) => ({
+            type: 'Feature',
+            geometry: { type: 'Point', coordinates: [0, 0] },
+            properties: {
+                [CATEGORY_TYPE]: searchTypeMap.GPX_TRACK,
+                [CATEGORY_NAME]: f.name,
+            },
+        }));
 }
 
 /**
